Refactored GameTable
