+++ src/features/mode/san/table/SearchGamesTable.js
@@ -10,7 +10,7 @@ import {
   TableRow
 } from '@mui/material';
 import * as modeConst from 'features/mode/modeConst';
-import * as sanMode from 'features/mode/sanModeSlice';
+import * as panel from 'features/panel/panelSlice';
 import * as variantConst from 'features/mode/variantConst';
 import * as nav from 'features/nav/navSlice';
 import Ws from 'features/ws/Ws';
@@ -56,19 +56,16 @@ const SearchGamesTable = ({props}) => {
   const handleLoad = (item) => {
     multiAction.initGui(dispatch);
     dispatch(nav.setDatabase());
-    dispatch(sanMode.panelTable({
-      open: true,
-      game: {
-        Event: item.Event,
-        Site: item.Site,
-        Date: item.Date,
-        White: item.White,
-        Black: item.Black,
-        'White ELO': item.WhiteElo,
-        'Black ELO': item.BlackElo,
-        Result: item.Result,
-        ECO: item.ECO
-      }
+    dispatch(panel.gameMetadataTable({
+      Event: item.Event,
+      Site: item.Site,
+      Date: item.Date,
+      White: item.White,
+      Black: item.Black,
+      'White ELO': item.WhiteElo,
+      'Black ELO': item.BlackElo,
+      Result: item.Result,
+      ECO: item.ECO
     }));
     const settings = {
       movetext: item.movetext
