+++ src/features/panel/MovesTable.js
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { Table, TableBody, TableCell, TableContainer, TableRow } from "@mui/material";
 import Movetext from 'common/Movetext.js';
-import GameTable from 'features/mode/san/table/GameTable';
+import GameMetadataTable from 'features/panel/GameMetadataTable';
 import * as panel from 'features/panel/panelSlice';
 import Ws from 'features/ws/Ws';
 
@@ -100,7 +100,7 @@ const MovesTable = ({props}) => {
       <TableContainer className="noTextSelection" sx={styles.table}>
         <Table stickyHeader size="small" aria-label="Movetext">
           <TableBody>
-            <GameTable />
+            <GameMetadataTable />
             {moves()}
           </TableBody>
         </Table>
