+++ src/features/panel/rav/RavMovesTable.js
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { Table, TableBody, TableCell, TableContainer, TableRow } from "@mui/material";
 import Movetext from 'common/Movetext.js';
-import GameTable from 'features/mode/rav/table/GameTable';
+import GameMetadataTable from 'features/panel/GameMetadataTable';
 import * as panel from 'features/panel/panelSlice';
 import Ws from 'features/ws/Ws';
 
@@ -132,7 +132,7 @@ const RavMovesTable = ({props}) => {
       <TableContainer className="noTextSelection" sx={styles.table}>
         <Table stickyHeader size="small" aria-label="Movetext">
           <TableBody>
-            <GameTable />
+            <GameMetadataTable />
             {description()}
             {moves()}
           </TableBody>
